+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2022-01-12
+### Changed
+- Message controller
+- Event resource wrapping
+- Event handle `last result`
+### Added
+- AlpineJs Support
+
 ## [1.0.0] - 2022-01-09
 ### Added
 - License mit.